[Refactor] Extract metadata generators from EtdMetadata
