+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-01-16 11:20:03 -0600 using RuboCop version 0.77.0.
+# on 2020-02-21 13:57:26 -0600 using RuboCop version 0.77.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -17,11 +17,6 @@ Lint/DeprecatedClassMethods:
   Exclude:
     - 'Rakefile'
 
-# Offense count: 1
-Lint/RescueException:
-  Exclude:
-    - 'lib/robots/dor_repo/accession/embargo_release.rb'
-
 # Offense count: 2
 # Configuration parameters: AllowComments.
 Lint/SuppressedException:
@@ -35,7 +30,7 @@ Lint/UnderscorePrefixedVariableName:
   Exclude:
     - 'config/boot.rb'
 
-# Offense count: 25
+# Offense count: 22
 Metrics/AbcSize:
   Max: 98
 
@@ -43,33 +38,28 @@ Metrics/AbcSize:
 # Configuration parameters: CountComments, ExcludedMethods.
 # ExcludedMethods: refine
 Metrics/BlockLength:
-  Max: 67
+  Max: 63
 
 # Offense count: 2
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
   Max: 158
 
-# Offense count: 8
+# Offense count: 6
 Metrics/CyclomaticComplexity:
   Max: 10
 
-# Offense count: 30
+# Offense count: 27
 # Configuration parameters: CountComments, ExcludedMethods.
 Metrics/MethodLength:
-  Max: 70
-
-# Offense count: 1
-# Configuration parameters: CountComments.
-Metrics/ModuleLength:
-  Max: 178
+  Max: 66
 
 # Offense count: 1
 # Configuration parameters: CountKeywordArgs.
 Metrics/ParameterLists:
   Max: 6
 
-# Offense count: 6
+# Offense count: 5
 Metrics/PerceivedComplexity:
   Max: 10
 
@@ -81,18 +71,9 @@ Naming/HeredocDelimiterNaming:
     - 'lib/technical_metadata_service.rb'
     - 'spec/lib/technical_metadata_service_spec.rb'
 
-# Offense count: 4
-# Configuration parameters: EnforcedStyle.
-# SupportedStyles: snake_case, camelCase
-Naming/VariableName:
-  Exclude:
-    - 'lib/robots/dor_repo/accession/end_accession.rb'
-
-# Offense count: 11
+# Offense count: 7
 RSpec/AnyInstance:
   Exclude:
-    - 'spec/lib/datastream_builder_spec.rb'
-    - 'spec/robots/accession/embargo_release_spec.rb'
     - 'spec/robots/assembly/jp2_create_spec.rb'
 
 # Offense count: 1
@@ -103,31 +84,23 @@ RSpec/BeforeAfterAll:
     - 'spec/support/**/*.rb'
     - 'spec/lib/technical_metadata_service_spec.rb'
 
-# Offense count: 19
+# Offense count: 16
 # Configuration parameters: Prefixes.
 # Prefixes: when, with, without
 RSpec/ContextWording:
   Exclude:
     - 'spec/robots/accession/content_metadata_spec.rb'
-    - 'spec/robots/accession/embargo_release_spec.rb'
-    - 'spec/robots/accession/provenance_metadata_spec.rb'
     - 'spec/robots/accession/technical_metadata_spec.rb'
     - 'spec/robots/assembly/content_metadata_create_spec.rb'
     - 'spec/robots/assembly/jp2_create_spec.rb'
 
-# Offense count: 1
-RSpec/DescribeClass:
-  Exclude:
-    - 'spec/robots/accession/embargo_release_spec.rb'
-
-# Offense count: 28
+# Offense count: 26
 # Configuration parameters: Max.
 RSpec/ExampleLength:
   Exclude:
+    - 'spec/lib/dor/etd/content_metadata_generator_spec.rb'
     - 'spec/lib/dor/etd/rights_metadata_generator_spec.rb'
     - 'spec/lib/technical_metadata_service_spec.rb'
-    - 'spec/models/etd_spec.rb'
-    - 'spec/robots/accession/embargo_release_spec.rb'
     - 'spec/robots/accession/provenance_metadata_spec.rb'
     - 'spec/robots/accession/shelve_spec.rb'
     - 'spec/robots/assembly/exif_collect_spec.rb'
@@ -135,35 +108,24 @@ RSpec/ExampleLength:
     - 'spec/robots/etd_submit/binder_batch_transfer_spec.rb'
     - 'spec/robots/goobi_notify_spec.rb'
 
-# Offense count: 1
-RSpec/ExpectInHook:
-  Exclude:
-    - 'spec/robots/accession/embargo_release_spec.rb'
-
 # Offense count: 27
 # Configuration parameters: CustomTransform, IgnoreMethods.
 RSpec/FilePath:
   Enabled: false
 
-# Offense count: 55
+# Offense count: 24
 # Configuration parameters: AssignmentOnly.
 RSpec/InstanceVariable:
   Exclude:
-    - 'spec/lib/datastream_builder_spec.rb'
-    - 'spec/lib/dor/release/item_spec.rb'
     - 'spec/lib/technical_metadata_service_spec.rb'
     - 'spec/robots/etd_submit/submit_marc_spec.rb'
-    - 'spec/robots/release/release_members_spec.rb'
 
-# Offense count: 63
+# Offense count: 45
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: have_received, receive
 RSpec/MessageSpies:
   Exclude:
-    - 'spec/lib/dor/release/item_spec.rb'
     - 'spec/lib/technical_metadata_service_spec.rb'
-    - 'spec/robots/accession/embargo_release_spec.rb'
-    - 'spec/robots/accession/technical_metadata_spec.rb'
     - 'spec/robots/assembly/checksum_compute_spec.rb'
     - 'spec/robots/assembly/content_metadata_create_spec.rb'
     - 'spec/robots/assembly/exif_collect_spec.rb'
@@ -171,12 +133,12 @@ RSpec/MessageSpies:
     - 'spec/robots/etd_submit/submit_marc_spec.rb'
     - 'spec/robots/release/release_members_spec.rb'
 
-# Offense count: 54
+# Offense count: 40
 # Configuration parameters: AggregateFailuresByDefault.
 RSpec/MultipleExpectations:
   Max: 60
 
-# Offense count: 21
+# Offense count: 13
 RSpec/NestedGroups:
   Max: 6
 
@@ -185,31 +147,23 @@ RSpec/ScatteredLet:
   Exclude:
     - 'spec/robots/assembly/jp2_create_spec.rb'
 
-# Offense count: 2
-RSpec/ScatteredSetup:
-  Exclude:
-    - 'spec/robots/accession/embargo_release_spec.rb'
-
-# Offense count: 1
+# Offense count: 6
 RSpec/SubjectStub:
   Exclude:
     - 'spec/robots/etd_submit/other_metadata_spec.rb'
 
-# Offense count: 24
+# Offense count: 19
 # Configuration parameters: IgnoreNameless, IgnoreSymbolicNames.
 RSpec/VerifiedDoubles:
   Exclude:
-    - 'spec/lib/technical_metadata_service_spec.rb'
-    - 'spec/models/etd_spec.rb'
-    - 'spec/robots/accession/provenance_metadata_spec.rb'
     - 'spec/robots/assembly/exif_collect_spec.rb'
     - 'spec/robots/etd_submit/submit_marc_spec.rb'
 
-# Offense count: 31
+# Offense count: 27
 Style/Documentation:
   Enabled: false
 
-# Offense count: 652
+# Offense count: 576
 # Cop supports --auto-correct.
 # Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
